Ruby: fix: fragment files were not always sorted
Not sure how it is I haven’t seen this before, but a few minutes ago I
ran this lesson and got fragment 2 first. This is very bad. This might
fix it — seems to, but I cannot yet be 100% sure. It’s at least a decent
try, and it documents the _intent_ that these files should be sorted.