Reorganize Ruby trees
As Sasha pointed out in #29, the old structure didn’t make readily
apparent which files were shared across contexts and which were not, and
required ugly, brittle hard-coded lists of file-paths in the start
scripts.

So now I’ve created dirs named `shared` at a few different levels in the
tree, and reorganized the tree so there’s a useful context for each of
those dirs, and updated the scripts accordingly.

Since this moves the location of the ruby and jruby courses (the former
contents of ruby have been moved into a new subdir of ruby named cruby,
and jruby has been moved next to cruby) I had to update the instructions
in the README, and as long as I was in there I made a few overdue
updates.

Thanks Sasha for the feedback!