Fix dir structure for CRuby Unit 2 (#36)
In 6973d827 I reorganized Ruby trees, but that branch was developed in
parallel with the branch wherein this unit was added (with just lesson
one so far). So somehow (shockingly) I missed some of the changes.

I’m not *entirely* sure how I missed this, but I do think I know why
this lesson has been working, even thought it didn’t *seem* to have
a docker-compose.yml — it’s because it *does* have a docker-compose.yml,
but itw was hidden in my editor because of the rules in the closest
.gitignore file.

Also, I’m pretty sure that after 6973d827 this lesson did actually
break, because `start.sh` tries to copy files from `../shared/*` — but
that dir didn’t exist until *now*. So it appears I didn’t test all the
lessons as part of that commit/branch. Which underscores the point that
@sgerrand made in #25 — that it’s really time for this project to have a
decent set of automated tests.