Switch start+stop scripts from hard-coded to dynamic project names
In the prior commit I added a test harness; one of the things that test
harness needs to do is to start/stop/etc the networks described in each
unit’s docker-compose.yml file. In order to do that consistently with
the lesson runner scripts (start.sh) which specify a project name rather
than going with the default, I have a few lines in test_all_units.sh that
programmatically construct a project name that’s consistent with how they’re
specified in each start.sh.

This worried me, though, because there was now a tight coupling between
the hard-coded project names in the start scripts and the programmatic logic
in the test harness. This seemed brittle to me. Hence this change, so that the
same logic is used everywhere.

This seems like a decent solution to the brittleness problem, but this is also
repetitive, redundant, and verbose. So it’s probably just about time to apply
some DRY and KISS to this situation.