Give docker-compose networks better names
I started working on these tutorials with a lesson-centric approach, but
I’ve since realized that WRT the infrastructure for the lessons, it’s
better to take a unit-centric approach. The lessons can still be run
mostly independently, or at least, in singular sessions — and that’s
important for pedagogical reasons — but they’ve turned out to be slightly
more coupled than I had originally anticipated, which I’m OK with.