Start: simplify + streamline Docker infrastructure
Starting in this commit with CRuby Unit 1, but that’s really only to prevent
this commit from being too sprawling to effectively review. Subsequent commits
should apply pretty much the same set of changes to the remaining units, and
adapt the test harness to these simplifications.

As I mentioned in the parent commit:

> I started working on these tutorials with a lesson-centric approach, but I’ve
> since realized that WRT the infrastructure for the lessons, it’s better to
> take a unit-centric approach. The lessons can still be run mostly
> independently, or at least, in singular sessions — and that’s important for
> pedagogical reasons — but they’ve turned out to be slightly more coupled than
> I had originally anticipated, which I’m OK with.

What I meant by “infrastructure” there is the Docker infrastructure used by the
units and lessons. The changes herein take advantage of this realization to
switch from each *lesson* being its own Docker context — which forced me to
construct a project name and supply it to every docker-compose command — to each
*unit* being the Docker context.

This means that the docker-compose commands no longer need to be passed a
project name, because the default name is now sufficient, because the docker
commands are now all always running from the same dir/context.

It also means we don’t need to copy the “shared” files for each unit from a
`shared` dir into each lesson dir — they’re now straightforwardly located in the
root of each unit dir and hence each docker container for each “unit” service
(as defined by `docker-compose.yml` and `Dockerfile`).