CRuby Unit 2: remove pre+post processing of hashes
I had wanted to use “pretty” and (more importantly) idiomatic hashes as
the data we’d serialize and deserialize via Avro, so I started with
hashes using snake_case keywords as keys, as is idiomatic in Ruby. But
it seems that the Ruby avro lib requires string keys.

I had also written the schema for the avro record with UpperCamelCase
field names, because — I’m not sure, but I think I was thinking that
that was the idiom for Avro. (Now that I think about it, I don’t think
Avro really cares — organization/system idioms are probably more
important than any idioms common in the community of Avro users. (Except
one thing: I’ve found (with some effort) that it’s a bad idea to use
spaces in record field names. The Ruby implementation of Avro supports
it, but the Java implementation does not.))

To deal with these incompatibilities between the pretty/idiomatic Ruby
hashes and the way I had written the Avro schemata, I had included these
pre- and post-processing steps, to transform the hashes into a form
that’d be compatible with the schemata, before serialization, and to
then transform the hashes deserialized from Avro back into the
pretty/idiomatic Ruby format.

Then I was working on lesson 3, and it became clear that these pre/post
processing steps were distracting, bloating, unnecessary
ceremony/plumbing that were detracting from the lessons.

Hence this change, which revises the schemata so they’re more compatible
with the input hashes, and reduces the pre/post-processing to
transforming the hash keys from symbols to strings and then back again.
I find this much less obtrusive.

I also changed the way in which the schemata are included in the code;
instead of a string in a heredoc, I changed them to Ruby hashes that are
then serialized to JSON — mainly just to make the schemata more readable
as the user is hopefully in the groove reading Ruby, and there’s really
no need to interrupt that in order to make them read JSON. The hashes
are also output with colorized syntax, so that should make them even
easier to read.