CRuby: drastically simplify basic record consumption
I realized that instead of client#each_message, which blocks and
therefore has necessitated all this thread ceremony around it, we can
use client#fetch_messages, which does not block, and thus we can
eliminate all the ceremony related to using a thread.

By “basic” I mean without using the Consumer Groups feature; it seems
the Consumer class of ruby-kafka (which does use that feature, as
opposed to the Client class which does not) doesn’t have a non-blocking
method for retrieving records.