CRuby Unit 2: Minor UX Tweak
Return (and therefore print to the terminal) the record values rather
than the records themselves, as the former just prints out way too much
stuff, and it’s not useful stuff. In these cases we really just want the
user to see that the records were consumed and that their values are
serialized Avro values.