[JRuby U1»L4] Use automatic serialization + deserialization
Just after I did the initial implementation of this lesson I realized
that I wasn’t doing serialization + deserialization the way that it’s
intended to be done with the Java Kafka clients. The way it’s supposed
to be done is that you supply them with Serializer and Deserializer
objects and the clients then use those objects to automatically
serialize and deserialize keys and values when producing and consuming
records, so that, among other advantages, the code that implements the
producing and consuming can be cleaner and simpler.

Thankfully JRuby makes it super easy to implement a Java interface — no
casting required. (And unlike in Java, we don’t even have to implement
all the methods in the interface, which is very nice.) I think this
improved implementation does a good job of demonstrating how smooth and
seamless it really is to use Java libraries from Ruby code using JRuby.